scripts/requirements.txt
pycryptodomex==3.22.0
